Extract CI signal helpers to keep github-merge.lib.mjs under 1500 lines
The fix in commit 2e2d9e62 added getWorkflowRunJobsCount (44 lines) which
pushed src/github-merge.lib.mjs from 1496 to 1540 lines, tripping the
check-file-line-limits CI gate (limit: 1500).

Move getCommitDate, checkPreviousPRCommitsHadCI, and checkWorkflowsHavePRTriggers
into a new src/github-merge-ci-signals.lib.mjs (their docs already group them
as the issue #1480 / #1503 "additional CI signals"), and re-export from
github-merge.lib.mjs so existing named and default-object imports keep working.

Result: github-merge.lib.mjs drops to 1379 lines; new file is 182 lines.
No behavior change.
